Refactor: Rename `lib/layouts` -> `lib/pages`
